mi: move ValidateRec to mi_priv.h and rename to MiValidateRec
