+++ hw/kdrive/src/kxv.h
@@ -42,7 +42,6 @@ of the copyright holder.
 #include "regionstr.h"
 #include "windowstr.h"
 #include "pixmapstr.h"
-#include "mivalidate.h"
 #include "validate.h"
 #include "resource.h"
 #include "gcstruct.h"
