+++ mi/mioverlay.c
@@ -18,7 +18,6 @@
 #include "gcstruct.h"
 #include "regionstr.h"
 #include "privates.h"
-#include "mivalidate.h"
 #include "mioverlay.h"
 #include "migc.h"
