+++ mi/miwindow.c
@@ -59,7 +59,6 @@ SOFTWARE.
 #include "windowstr.h"
 #include "scrnintstr.h"
 #include "pixmapstr.h"
-#include "mivalidate.h"
 #include "inputstr.h"
 
 void
@@ -123,11 +122,13 @@ miClearToBackground(WindowPtr pWin,
 void
 miMarkWindow(WindowPtr pWin)
 {
-    ValidatePtr val;
-
     if (pWin->valdata)
         return;
-    val = (ValidatePtr) XNFalloc(sizeof(ValidateRec));
+
+    ValidatePtr val = (ValidatePtr) calloc(1, sizeof(MiValidateRec));
+    if (!val)
+        return;
+
     val->before.oldAbsCorner.x = pWin->drawable.x;
     val->before.oldAbsCorner.y = pWin->drawable.y;
     val->before.borderVisible = NullRegion;
